+++ logic/src/main/java/application/Repository/RoleRepository.java
@@ -3,8 +3,6 @@ package application.Repository;
 import application.models.Role;
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.List;
-
-public interface RoleRepository extends CrudRepository<Role, Integer>{
-    Role findById(int roleId);
+public interface RoleRepository extends CrudRepository<Role, Integer> {
+    Role findById(long roleId);
 }
