+++ logic/src/main/java/application/models/Role.java
@@ -18,7 +18,7 @@ import javax.persistence.Id;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
     private String name;
-    
+
 }
